% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_series.R
\name{delete_series}
\alias{delete_series}
\title{Deletando aberturas/transformações para a FS}
\usage{
delete_series(indicator = as.character(), token = as.character())
}
\description{
Deletando aberturas/transformações para a FS
}
\examples{
\dontrun{
}

}
\author{
Luiz Paulo Tavares Gonçalves
}
