% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_body.R
\name{get.body}
\alias{get.body}
\title{Estrutura Body da Api}
\usage{
get.body()
}
\value{
O retorno é um JSON com o body que precisa ser preenchido para a criação de novos indicadores
}
\description{
A função retorna o body da api (verbo POST) utilizado na criação de novos indicadores
}
\examples{

\dontrun{
body_fill = fsApi::get.body()
}

}
\author{
Luiz Paulo Tavares Gonçalves
}
