% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_master_node.R
\name{get.master_node}
\alias{get.master_node}
\title{Consultar nó mestre, isto é, o primeiro nó}
\usage{
get.master_node(token = as.character())
}
\arguments{
\item{token}{representa o token de acesso a FS;}
}
\value{
O retorno é um tibble com todos os nós da primeira ramificação
}
\description{
A função busca consultar os nós (mestres) da árvore da FS
}
\details{
O token necessariamente precisa está no formato de string
}
\examples{
\dontrun{

fs_master_node = fsApi::get.master_node(token = token_prod)

}

}
\author{
Luiz Paulo Tavares Gonçalves
}
