% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_indicator.R
\name{post.indicator}
\alias{post.indicator}
\title{Enviar indicadores para a FS}
\usage{
post.indicator(
  name_en = as.character(),
  name_pt = as.character(),
  short_en = as.character(),
  short_pt = as.character(),
  source_en = as.character(),
  source_pt = as.character(),
  description_en = as.character(),
  description_pt = as.character(),
  description_full_en = as.character(),
  description_full_pt = as.character(),
  unit = as.character(),
  country = as.character(),
  sector = as.character(),
  acess_group = as.character(),
  node_en = as.character(),
  node_pt = as.character()
)
}
\arguments{
\item{name_en}{representa o nome do indicador em inglês;}

\item{name_pt}{representa o nome do indicador em português;}

\item{short_en}{representa o nome curto em inglês;}

\item{short_pt}{representa o nome curto em português;}

\item{source_en}{representa a fonte/origem dos dados em inglês;}

\item{source_pt}{representa a fonte/origem dos dados em português;}

\item{description_en}{representa a descrição do indicador em inglês;}

\item{description_pt}{representa a descrição do indicador em português;}

\item{description_full_en}{representa a descrição completa em inglês;}

\item{description_full_pt}{representa a descrição completa em português;}

\item{unit}{representa a unidade de medida do indicador;}

\item{country}{representa o país dos dados presente no indicador;}

\item{sector}{representa o setor no qual o indicador se insere;}

\item{acess_group}{representa o grupo no qual o indicador está;}

\item{id_node}{é um data.frame com os nodes e ids de cada node do indicador.}
}
\value{
O retorno é um dataframe com todos os indicadores criados/enviados para a FS
}
\description{
A função busca enviar/criar novos indicadores na FS
}
\details{
Caso já exista o indicador na FS, a função irá criar o mesmo indicador, porém com numeração diferente.
}
\examples{
\dontrun{

send_indicadores = post.indicator(name_en = "ARG - Exports of Goods",
                                 name_pt = "ARG - Exportações de Bens",
                                 short_en = "Export. Goods",
                                 short_pt = "Exp. Bens",
                                 source_en = "Indec",
                                 source_pt = "Indec",
                                 description_en = "....",
                                 description_pt = "...",
                                 description_full_en = "...",
                                 description_full_pt = "...",
                                 unit = "USD Milhões",
                                 country = "AR",
                                 sector = "BOP",
                                 acess_group = "Geral",
                                 node_en = "Brazil",
                                 node_pt = "Brasil")
}

}
\author{
Luiz Paulo Tavares Gonçalves
}
