% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_series.R
\name{post.series}
\alias{post.series}
\title{Enviar aberturas/transformações para a FS}
\usage{
post.series(
  indicator_code,
  region_code,
  transfs_code,
  units_en,
  units_pt,
  token
)
}
\arguments{
\item{indicator_code}{Indicador de 9 dígitos;}

\item{transfs_code}{Contém a região, agregação, transformação primária e secundária;}

\item{units_en}{A unidade de medida de cada série input em Inglês;}

\item{units_pt}{A unidade de medida de cada série input em Português;}

\item{token}{Token de acesso do usuário.}
}
\value{
O retorno é um dataframe com todos os sids criados/enviados para a FS
}
\description{
A função busca enviar as aberturas/transformações solicitadas pelo usuário para a FS
}
\examples{
\dontrun{
send_transf = fsApi::post.series(indicator_code = "BRBOP0001",
                                 region_code = c('AL2', 'GO2'),
                                 transfs_code = c("OOML", "OOMM"),
                                 units_en = c("teste1", "teste2"),
                                 units_pt = c("testea", "testeb")
                                 token = headers)
}

}
\author{
Luiz Paulo Tavares Gonçalves
}
